refactor custom field edit modal into its own component
