+++ packages/front-end/components/Features/FeaturesOverview.tsx
@@ -83,9 +83,8 @@ import usePermissionsUtil from "@/hooks/usePermissionsUtils";
 import CustomMarkdown from "@/components/Markdown/CustomMarkdown";
 import Markdown from "@/components/Markdown/Markdown";
 import EditFeatureDescriptionModal from "@/components/Features/EditFeatureDescriptionModal";
-import CustomFieldDisplay, {
-  CustomFieldDraftInfo,
-} from "@/components/CustomFields/CustomFieldDisplay";
+import CustomFieldDisplay from "@/components/CustomFields/CustomFieldDisplay";
+import { CustomFieldDraftInfo } from "@/components/CustomFields/CustomFieldEditModal";
 import { useCustomFields } from "@/hooks/useCustomFields";
 import { filterCustomFieldsForSectionAndProject } from "@/services/customFields";
 import { useLocalStorage } from "@/hooks/useLocalStorage";
